+++ backend-java/src/main/java/com/bytebites/service/IUserInfoService.java
@@ -1,0 +1,16 @@
+package com.bytebites.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import com.bytebites.entity.UserInfo;
+
+/**
+ * <p>
+ *  服务类
+ * </p>
+ *
+ * @author 虎哥
+ * @since 2021-12-24
+ */
+public interface IUserInfoService extends IService<UserInfo> {
+
+}