+++ backend-java/src/main/java/com/bytebites/utils/RefreshTokenIntercepter.java
@@ -1,0 +1,56 @@
+package com.bytebites.utils;
+
+import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.util.StrUtil;
+import com.bytebites.dto.UserDTO;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.lang.Nullable;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+public class RefreshTokenIntercepter implements HandlerInterceptor {
+
+
+    private StringRedisTemplate stringRedisTemplate;
+
+    public RefreshTokenIntercepter(StringRedisTemplate stringRedisTemplate) {
+        this.stringRedisTemplate = stringRedisTemplate;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        //1.獲取請求頭的token
+        String token = request.getHeader("authorization");
+        if (StrUtil.isBlank(token)) {
+            return true;
+        }
+        //2.基於token獲取redis中的用戶
+        String key = RedisConstants.LOGIN_USER_KEY + token;
+        Map<Object, Object> userMap = stringRedisTemplate.opsForHash().entries(key);
+
+        //3.判斷用戶是否存在
+        if (userMap.isEmpty()) {
+            return true;
+        }
+        // 5.將查詢到的HASH數據轉為UserDTO對象
+        UserDTO userDTO = BeanUtil.fillBeanWithMap(userMap, new UserDTO(), false);
+        //6.存在，保存用戶資訊到ThreadLocal
+        UserHolder.saveUser(userDTO);
+
+        //7.刷新token有效期
+        stringRedisTemplate.expire(key, RedisConstants.LOGIN_USER_TTL, TimeUnit.MINUTES);
+
+        //8. 放行
+        return true;
+    }
+
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, @Nullable Exception ex) throws Exception {
+        //移除用戶
+        UserHolder.removeUser();
+    }
+}