+++ backend-java/src/main/java/com/hmdp/enums/LockType.java
@@ -1,6 +1,0 @@
-package com.hmdp.enums;
-
-public enum LockType {
-    READ,
-    WRITE
-}